\name{get_example_regions}
\alias{get_example_regions}

\title{
return a set of regions for the examples and unit testing
}
\description{

return a set of regions for the examples and unit testing
}
\usage{
get_example_regions()

}

%\arguments{
%}
%\details{

%}
%\value{

 



%}
\author{
Daryl Waggott
}
\examples{
index <- get_example_regions()
}
\keyword{ ~kwd1 }
