\name{index2bed}
\alias{index2bed}

\title{
convert a region index into a bed file dataframe
}
\description{
convert a region index into a bed file dataframe

}
\usage{
index2bed(x)
}

\arguments{
  \item{x}{an index
}
}
%\details{
%}
%\value{
%}
\author{
Daryl Waggott
}
\examples{
#a.bed <- index2bed(x) 
}
\keyword{bedtools}
