\name{modifyList2}
\alias{modifyList2}
\title{
Interface to R's modifyList
}
\description{
Interface to R's modifyList
}
\usage{
modifyList2(x, val)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{a named list}
  \item{val}{a named list with components to be updated using x}
}
%\details{
%}
\value{
modified version of x
}
%\references{
%}
\author{
Daryl Waggott
}

\seealso{
modifyList
}

%\keyword{ }
