\name{strsplit2matrix}
\alias{strsplit2matrix}

\title{
split a vector of strings into tabular data
}
\description{
split a vector of strings into tabular data

}
\usage{
strsplit2matrix(x, split, fixed = FALSE, perl = FALSE)
}

\arguments{
  \item{x}{a character vector
}
  \item{split}{the character or regex to split on}
  \item{fixed}{fixed i.e. no regex}
  \item{perl}{per style}
}
%\details{
%}
%\value{
%}
\author{
Daryl Waggott
}
\examples{
# a.bed <- strSplitToMatrix(x) 
}
\keyword{bedtools}
