\name{table2venn}
\alias{table2venn}
\title{
Plot venn diagram
}
\description{
Plot venn diagram of regions intersect
}
\usage{
table2venn(x, var.names)
}
\arguments{
  \item{x}{intersect table of regions}
  \item{var.names}{names of the overlapping regions}
}
%\details{
%}
\value{
venn diagram input list
}
%\references{
%}
\author{
Daryl Waggott
}
\seealso{
bedr.plot.region
}

%\examples{
%}
\keyword{venn}
