\name{bed2index}
\alias{bed2index}

\title{
bed dataframe to index string
}
\description{

convert a dataframe in bed format to an index string
}
\usage{
bed2index(x, sort = TRUE)

}

\arguments{
  \item{x}{ a object region object or index}
  \item{sort}{ should the index be sorted}
}
%\details{
%}
%\value{
%}
\author{
Daryl Waggott
}
%\examples{
%}
\keyword{ bed }
