\name{bedr-package}
\alias{bedr-package}
\alias{bedr-package}
\docType{package}
\title{
A bedtools wrapper for working with genomic ranges in R
}
\description{
A bedtools wrapper that allows using a mix of internal R objects and external R files.  A number of convenience functions are provided for simplifying analysis workflows in R.
}
\details{
\tabular{ll}{
Package: \tab bedr\cr
Type: \tab Package\cr
Version: \tab 1.0.4\cr
Date: \tab 2016-09-19\cr
License: \tab GPL2\cr
}
}
\author{
Daryl Waggott\cr
Syed Haider
}
\keyword{ package }
