\name{createTmpBedFile}
\alias{createTmpBedFile}

\title{
output R objects as tmpfiles
}
\description{
output R objects as tmpfiles

}
\usage{
createTmpBedFile(x, name = "bedr", tmpDir = NULL)
}

\arguments{
  \item{x}{region object}
  \item{name}{a prefix for the tmp file.}
  \item{tmpDir}{where should the temp files be put}
}
%\details{
 
%}
%\value{
%}
\author{
Daryl Waggott
}
\examples{
# create tmp file 
}
\keyword{tmp}
