\name{get.example.regions}
\alias{get.example.regions}

\title{
return a set of regions for the examples and unit testing
}
\description{

return a set of regions for the examples and unit testing
}
\usage{
get.example.regions()

}

%\arguments{
%}
%\details{

%}
\value{
A list with three example regions
}
\author{
Daryl Waggott
}
\examples{
index <- get.example.regions()
}
\keyword{ ~kwd1 }
